// src/type/repayment_ticket.d.ts
export interface RepaymentTicketJson {
  feePriceValue: string;
  repayPriceValue: string;
  state: string;
  usedPriceValue: string;
  calculateTarget: string;
  toTeikiRouteSectionIndex: string;
  fromTeikiRouteSectionIndex: string;
  validityPeriod: string;
  payPriceValue: string;
  changeableSection: string;
}